requirements.txt
# API and HTTP libraries
requests>=2.28.1
httpx>=0.23.0

# Async support
asyncio>=3.4.3

# Database libraries
psycopg2-binary>=2.9.3

# Configuration and parsing
pyyaml>=6.0

# Scheduling
schedule>=1.1.0

# Date and time handling
python-dateutil>=2.8.2

# Testing
pytest>=7.0.0
pytest-mock>=3.7.0

# Development tools
black>=22.1.0
flake8>=4.0.1
isort>=5.10.1